Cloud_Computing/ML_Rest_API/requirements.txt
flask
opencv-python
opencv-contrib-python
tensorflow
pillow